+++ app/src/main/java/com/example/bookstats/network/ApiService.kt
@@ -9,23 +9,3 @@ interface ApiService {
     @GET("volumes")
     suspend fun getBookByISBN(@Query("q") isbn: String): Response<ApiResponse<BookResponse>>
 }
-
-data class ApiResponse<T>(
-    val items: List<T>
-)
-
-data class BookResponse(
-    val volumeInfo: VolumeInfo
-)
-
-data class VolumeInfo(
-    val title: String,
-    val authors: List<String>,
-    val pageCount: Int,
-    val imageLinks: ImageLinks?
-)
-
-data class ImageLinks(
-    val smallThumbnail: String?,
-    val thumbnail: String?
-)
