+++ app/src/main/res/layout/fragment_book_creation.xml
@@ -10,7 +10,7 @@
         android:id="@+id/book_image_container"
         android:layout_width="200dp"
         android:layout_height="250dp"
-        android:layout_marginTop="60dp"
+        android:layout_marginTop="50dp"
         app:cardCornerRadius="16dp"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
@@ -19,7 +19,7 @@
         <ImageView
             android:id="@+id/book_image"
             android:layout_width="200dp"
-            android:layout_height="250dp"
+            android:layout_height="260dp"
             android:contentDescription="@string/book_desc_image"
             android:scaleType="centerCrop"
             android:src="@drawable/add_book_image"
@@ -28,25 +28,6 @@
             app:layout_constraintTop_toTopOf="parent" />
     </androidx.cardview.widget.CardView>
 
-    <com.google.android.material.textfield.TextInputLayout
-        android:id="@+id/isbn_code"
-        style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="24dp"
-        android:layout_marginTop="40dp"
-        android:layout_marginEnd="24dp"
-        app:layout_constraintBottom_toTopOf="@id/book_image_container"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent">
-
-        <com.google.android.material.textfield.TextInputEditText
-            android:id="@+id/isbn_code_editText"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:hint="@string/temp" />
-    </com.google.android.material.textfield.TextInputLayout>
-
     <Button
         android:id="@+id/temp_isbn"
         android:layout_width="wrap_content"
@@ -54,7 +35,9 @@
         android:layout_marginTop="60dp"
         android:text="@string/temp"
         app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        app:layout_constraintTop_toTopOf="parent"
+        android:visibility="gone"
+        />
 
     <com.google.android.material.textfield.TextInputLayout
         android:id="@+id/book_name"
